actual replacement of the previous logic
